# Improve performance of InputSplitter