@rpep 2017-11-14
I should not that I don't habitually write 500 line cells; I only noticed this because I'm doing some autogeneration work!