@Carreau 2017-11-14
can you make `indent_space` a property (read_only) to change less the API if you are concern ?