@takluyver 2017-11-15
Thanks @Carreau @willingc - I have updated this following your suggestions.