@takluyver 2017-11-15
What do we think about backporting this to 5.x? I think I lean slightly towards doing so, but I don't have a strong preference.