@Carreau 2017-11-15
Let's see how hard it is. 

@meeseeksdev backport to 5.x