@takluyver 2017-11-15
OK, I've marked this as 6.3. It was like that for years already, so I doubt many people are hitting it.