@takluyver 2017-11-15
It might be a neat feature for meeseeks if it could display the conflict - or possibly a summary of it if it's big.