@Carreau 2017-11-15
Yeah, I though about that, the secret command is `git diff --diff-filter=U` after the fail cherry-pick. I might do that at some point.