@codypiersall 2018-01-29
@takluyver I ran into this issue today, and this fixed it.  Thank you!  Now it doesn't take forever to load the UI for a project I am working on!